twikit_scraper/requirements.txt
twikit 
python-dotenv 
rich 
fastapi
uvicorn[standard] 